+++ frontend/src/routes/+layout.svelte
@@ -10,14 +10,8 @@
 	import { m } from '$lib/paraglide/messages.js';
 	import { getLocale, locales } from '$lib/paraglide/runtime';
 	import { callErrorToast } from '$lib/toast';
-	import {
-		clickOutside,
-		currentVersion,
-		get_prefs,
-		isFormDirty,
-		set_lang,
-		Version
-	} from '$lib/ui';
+	import { clickOutside, get_prefs, isFormDirty, set_lang } from '$lib/ui';
+	import { currentVersion, Version } from '$lib/enums/version';
 	import { Toaster } from 'svelte-sonner';
 	import '../app.css';
